Add new robust initial wrapping guess for arm muscles
This implements new initial guess for the arm wrapping muscles in the arms.
It uses the SPLine.InitWrapPosVecArr to set initial wrapping instead of the
old InitWrapPosVectors.

Besides being more robust when the model starts in extreme postures, it has
the benefit that the model does not look like a pin-cushion when running
initial positions.