Simplification of scapula reactions
removed slider segments usage for scapula/thorax contact and replaced it with pointsurface measure this removes two segs for each side.
fixed symmetry issue on deltoid rake so that p1 and p2 are symmetric ... it should not have changed the results except for non linear morphed models. p1 and p2 only used to find a radius.