Update hip positions for old Leg
This fixes the hip joint positions in the pelivis segment when using
the old leg model. The points were not updated when the New pelvis were
introduced. The Pelvis hip joint center are chosen be the center of the
hip socket, similar to what is used on TLEM1.

This entails that we also update the femural hip center of the old leg,
so the two center align correctly. The femural head center was not
 located at visual center of the femural head.