Update the hamstring and gastroc wrapping surface
This improves the wrrapping surface and via points of the gastroc and hamstring muscles around the knee. The fix ensures that the muscles have a moment arm when the knee is fully extended. The wrapping surface has no effect when the knee starts to flex. The wrapping surface is constructed fromt the TLEM2 MRI scans.

The fix was necessary, as the muscles had almost no moment arm when the knee was fully extended. The consequence was high activty at the end of the swing phase where the muscles decelerate the lower leg. In LLJ this caused a third peak in the hip contact forces, because the biarticular hamsrtrings were doing to much.