Restructure BM parameters for enabling TLEM2 ModelSetup.any
This adds TLEM2 as one of the models which selected with ``#BM_LEG_RIGHT`` and ``#BM_LEG_LEFT``, thus dropping the ``TLEM_VERSION`` statement we have used until now.

#define BM_LEG_LEFT _LEG_MODEL_TLEM2_

It also renames the ``TLEM_MORPHOLOGY`` to ``BM_LEG_MORPHOLOGY`` and creates a ``BM_LEG_DATASET`` parameter which can control which dataset folder should be set (e.g. "TLEM1.1"/"TLEM2.1").

It also introduces a ``BM_LEG_MODEL`` parameter which holds the default value for ``BM_LEG_RIGHT`` and ``BM_LEG_LEFT``. Thus, it can be used to set both legs simultaneously.

``BM_LEG_MODEL`` has an extra advantage, it can be used when checking which Model type is used in the model.

E.g. checking if anytype of TLEM model is in used would otherwise look like this:

#if (BM_LEG_RIGHT == _LEG_MODEL_TLEM2_) | (BM_LEG_LEFT == _LEG_MODEL_TLEM2_) | (BM_LEG_RIGHT == _LEG_MODEL_TLEM_) | (BM_LEG_LEFT == _LEG_MODEL_TLEM_)

This can now be:

#if BM_LEG_MODEL == _LEG_MODEL_TLEM2_ | BM_LEG_MODEL == _LEG_MODEL_TLEM_