Refactor Scaling system in HumanModel
This commit does the following to the HumanModel scaling system:

Restructures the AnyMan files, with a more consistent structure. Both new variable and new folder names.

An AnyMan files is now always generated even if it haven't been set through 'BM_SCALING_ANTHRO_FILE'. The folde with Anthropometic values are generated using a class_template which allows the values to be overridden by an application.

It also adds new Scaling law, _SCALING_XYZ_, which a sets the scaling segments in all three directions.