Restructure Custom Scaling
This removes the need for nested structure for the Custom Scaling.

Now if custom scaling is used the DefaultScaling folder is placed in a
folder 'DefaultScaling' in a folder parallel to the 'Scaling' folder create
by "Scaling/ScalingCustom.any".

This means that relative links in the different scaling laws will always work
because files are always included on the same level.