Add correction to Thorax Anatomical frame
After the changes to the Pelvis anatomical frames, the thorax position no longer corresponds
to neutral when the model is driven to Pelvis-Thoarax_extension of 0 degrees.

This changes corrects the anatomical frame so the hip joint center aligns with
the skull C1C0 joint in a neutral position. It also corrects the ij/iji nodes from the arms
to ensure the hangs straight down in a neutral position.