Fix old legs muscle insertions on the new Pelvis
The update to the Trunk pelvis in AMMR 2, corrupted the muscles
insertions for the old leg. The fix is to align muscle insertions as best
as possible with the new pelvis. The muscle insertions are still not
perfectly aligned with new Pelvis. This is impossible to do without
some form of interface morphing. However, interface morphing would
also substantially change the output of the leg model.