Fix Teres Minor wrapping convergence problems
Moved the point controlling the wrapping surface for the humerus upwards to
ensure that some muscles do not wrap on the caps. Also added cap ratio of 1
to the cylinder to improve wrapping

Teres Minor insertions are adjust a by two millimeters closer to the
bone surfaces, and the abs tolerance for the wrapping is lowered.