Fix regression in GRF contact area visualization
This was recently changed to support Base frames different from global reference frame.
However, since global reference frames do not have a ARel/r but Axes/Origin the code
would not work in the original case.

This fixes the problem by introducing a AnyFunTransform3DLinRef to handle transformation.
The only downside is that this introduces a dependency on how the Global reference is defined.

The class template now has a optional argument which can be specified if the global reference
is not Main.EnvironmentModel.GlobalRef