Added an option to the ForcePlateAutoDetection for exclusive limb detection
Added an option to the ForcePlateAutoDetection for class template to make the limb1/2 contact detection
mutually exclusive. Setting the option ``EXLCLUSIVE_LIMB_CONTACT=ON`` will ensure that both legs cant
be in contact with the plate at the same time. This can prevent accidental contact detection
for the collateral leg in the swing phase.