Update MoCap filter warning no to trig on Standing reference files
This changes the logic in warning for tStart/tEnd being close to the start/end frames
of the c3d files. It no longer triggers when the file is very short or when the
the model is configured to run only a single time frame (e.g. with a standing reference trial)