Update wrist flexion definition in BodyModel.Interface
The PR fixes kinematic problems when driving specific wrist-flexion/abduction values.
The definition of wrist flexion in ``BodyModel.Interface`` now measures the
flexion between the radius and small artificial ``WristJointSeg`` segment
between the radius and the hand segment. This ensures that there is always
unique solution when driving the wrist flexion and abduction.