Change pushup model posture to something more realistic
This change gives a more realistic posture for the pushup model. It replaces
the wrist abduction driver with a driver that keeps the elbow joint at fixed
distance relative to the shoulder. Thus, effectively controls the type of
push up (e.g. elbows in/out).

The result is a posture which is more natural and overloads the muscles less.