Ensure acceleration in MoCap is correct
This PR increases the start/end offset when running inverse dynamics.
The offset is currently 2 frames which are skipped at the start/end
of inverse dynamics simulations. But the accelerations of 4th order
b-spline are only stable on the fourth frame, and hence the offset
must be three frames instead of two.