Improve wrapping for Teres Minor.
The Teres Minor muscle now has its own wrapping cylinder on the humerus. as well as wrapping
cylinder in the scapula reference frame. Both cylinders meet at the glenohumeral head and prevent the teres minor
from sliding on to the wrong side of the humerus wrapping cylinder in extreme range of motions.