TLEM1. Ensure Satorius branches follow the same line of action.
This is a provisional fix for a long standing error in the TLEM1 model. The orginal TLEM1 paper intended the Satorius muscle to be implemented as two muscles in series. This was never implemented that way in the AnyBody implementation. This resulted in the volume of the Satorius beeing incoorect. It also mean the SatoriusDistal muscle in TLEM1 had the wrong line of action. This fixes the line of actions so SatoriusDistal/proximal share the same via points.