Fix Satorius muscle volume.
The volume was dobule the actual values because the value reported the orignal TLEM1 paper were intended for an implementation where the Satorius was implemented as two muscle in series. It was never implemented that way in AnyBody so the muscle volume wrong. This fixes both the TLEM1 where Satorius has two strands in parallel, and TLEM2 where Satorius is a single element muscle.