Improve the shoulder rhythm
Scapula position when using the shoulder rhythm is now determined by the Conoideum ligament driver. Similar to how scapula is
driven when the shoulder rhythm is not enabled.

The rhythm is now controlled based on the 'plane of elevation' and 'shoulder elevation' from the 'ThoraxHumerus' measures in the interface folder. This tracks the original specification of the shoulder rhythm in "De Groot, J. H. The shoulder: a kinematic and dynamic analysis of motion and loading. (1998)".