Fix scaling bug in _SCALING_UNIFORM_ and _SCALING_LENGTHMASSFAT_
This fixes a bug with using _SCALING_UNIFORM_ and _SCALING_LENGTHMASSFAT_
and only specifing the body height as input.

The segment length were set to slightly smaller value, which caused the total
body height to be less than what used as input.

The bug occured since a body for the default segments length in AnyMan files
was assumed to be 1.8 m instead of the correct 1.75.

This fix alligns the results with `__SCALING_XYZ_` (used int the MoCap models) and
`_SCALING_STANDARD_` (50% persentile) which were both correct.