Improve muscles for trunk rotation (#584)
Improve implementation of internal/external obliques and latissimus dorsi muscles.

This improves the strength for trunk external rotation. More branches are added to the internal obliques and the geometry is update to better match anatomical text books. Likewise, the latissimus dorsi muscles is discretized into more branches so it has branch inserting on every vertebra.