Update FemoralTorsionInclude.any
Added an #if statement to visualize the morphing points quickly.

Corrected some names for the visualization of the points in the joints. Now in the unscaled model it shows really well that the knee and the hip joint centers are not  translated at all, but it's only the geometry around them that gets morphed.