Additional image in the documentation + edit torsional values in the example
Added an image to the documentation to show the effect of the morphed torsion on muscle lines of action.  I'm not sure it is included correctly in the syntax of the doc.

Changed the default value in the example to 20. This value is considered within the normal healthy range, but still it's quite different from the baseline of the TLEM cadaver. -5 would have been retroversion. It's less common and also pathological.

Changed the reference value of the baseline model to 5.5 so it is consistent with the paper until a better way to calculate this is implemented.