Rewrite the femoral torsion code
This fixes a number of smaller issues which all comes from adding the femoral
torsion rotation in a coordinate system that is not completly alligned with
the anatomical reference frame. The anatomical frame is define based on
bony landmarks but we were rotating around a line between hip and functional
knee joint.

This changes also adds a way to measure the actual femoral torsion in the
model. Both with and and without scaling. The value before scaling can
then be used as a baseline when users add their target femoral torsion.

Finally this structures the code, so the right and left side shares the
same file. Minimizing the risk of diferences in the model.