In some cases use small pertubation to GlenohumeralAbduction
Show a warning and automatically apply a very small pertubation to the initial
orientation (Axes0) of the Humerus bone when the mannequin values of both GlenohumeralFlexion and GlenohumeralAbduction are initialized to zero.

This prevents probems with drivers which uses the "plane of elevation" measures for the ThoraxHumerus. Fx the shoulder rhythm.