Neutral scapula posture (#637)
* Update the neutral Scapula position and scapula sliding

The default neutral scapula position (medial) rotation have been updated
to a more realistic posture. Additionally, the node at scapula ts on which scapula
slides on the thorax have been moved to provide a more realistic clearance between
the scapula and the ellipsoid sliding surface representing thorax.

Thanks to Johanna Menze from the Univesity of Bern from updating the model.


* Minor tweak to LatissimusDorsi wrapping to prevent wrapping issues.

This is only an issue in 7.3.4. 7.3.5/7.4 are robust enough.

* Increase the thresshold for arm symmetry errors slightly

* Adjust start guess in flexed lift model.