Fix test cases for Knee simulator model (#798)
We updated how the class AnySurfCylinderFit is used in the {ref}KneeSimulator example <example_kneesimulator>. The deprecated way of using the AnySurfCylinderFit have been changed. Since AnySurfCylinderFit is now a sub class of AnyRefNode it should no longer be nested inside inside an other parent reference frame to provide the pos sRel and orientation ARel. This fixes a number of deprecated warnings which would have triggered with AMS 7.5.

* Update AnySurfEllipsoidFit to new recommended use

* Update test of Knee Simulator model

Some test are no longer present in AMS 7.5, and we fix the deprecated
way to use AnySurfFit

* Add changelog entry

* Add option to override pytest test argument

* Clean up the workflow file

* Add extra checkbox to control custom AMS

* Readd exception for test failure