Remove Marker Tracking  as initial condition for Inverse dynamics
We originally thought that was a good idea since it could handle special
case where the model started in 180 rotated position. But it seems to have
some drawbacks which Søren also discovered. So we fall back to the previous
implemetation of using standard initial positions.

We do however use the over-determinate solver for Initial Conditions. It seems
to be slightly more robust, and the lower speed doesn't matter for  a single
timestep with high kinematic tolerance.