Tlem foot fixes3 (#883)
Redefined the the ankle complex  and the knee-joint axis with respect to tibia based on the original CT scans of the TLEM 2 dataset. The ankle and subtalar joints are now based on the calcaneus, talus and tibia bones and how they fit together.  Previously the ankle complex were based on a unloaded functional trial performed during the cadaver dissection.

The new ankle and subtalar joint axis definition makes the bones fit much more nicely together. 

This update also changes the knee joint axis with respect to Tibia. This compensates for the 'non-neutral' position in which the cadaver was CT scanned.  Hence, the feet continue to point same direction in the neutral position. Essentially the tibia is rotated around its own long axis. This fix ensures that the patella tendon is straight. We believe this change eliminates many of the artifacts of the original TLEM 2 model caused by the  non-neutral position in which the cadaver was scanned. 

---------

Co-authored-by: Morten Enemark Lund <mel@anybodytech.com>