fix: Remove unused BM_EXPERIMENTAL statements (#78)
* Remove BM_TRUNK_EXPERIMENTAL_FLOATINGRIBS_KINETICS
The default is used, which is actuators at the end of floating ribs (floating ribs are rib 11 and 12).

* Remove BM_TRUNK_EXPERIMENTAL_RIBCAGE_REACTION
The default: Reaction in 6DOF sternum AVG driver + actuators in 3 directions in CC joints.

* Remove BM_EXPERIMENTAL_ERECTORSPINAE_VIANODES
Default: Lateral via nodes are placed on layers and medial ones stayed on vertebrae since they are close enough to the vertebrae. The L3 and L4 lateral via nodes are removed.

* Test BM_TRUNK_EXPERIMENTAL_RECTUS_ABDOMINIS

* fix bug: when buckle is used, lateral via nodes are placed on vertebrae as before

---------

Co-authored-by: hs <hs@anybodytech.com>