feat(MoCap): Add an AllDOF variable to the JointAngleOuput
Two new outputs are added to the JointAngleOutputs
folder in the MarkerTracking study in MoCap
models.

* `JointAngleOutputs.AllDoFs.Values`
* `JointAngleOutputs.AllDoFs.Names`

The variables contain values and names of all
variables which are saved when marker tracking
runs. They provide a different way to get the same
intermediate joint angle values which AnyBody
writes to files when running marker trakcing.